requirements.txt
novita-client
huggingface_hub
python-dotenv